Put some common function into util
