+++ Tools/test-connectivity.js
@@ -3,70 +3,20 @@
  */
 'use strict';
 
-var bluetoothctl = require('./lib/bluetoothctl.js');
-var util = require('./lib/util.js');
+var testConnectivity = require('./lib/test-connectivity.js');
+var util = require('./util.js');
 
 (function(mac) {
-  // the bluetootchctl only supports upper case
-  mac = (mac + '').toUpperCase();
-  // check whether the mac address is correct
-  if (!valid(mac)) {
-    usage();
-    return;
+  try {
+    testConnectivity(mac, connectSuccess, connectFail);
+  } catch (err) {
+    var errorMessage = err.message || err;
+    if (errorMessage.startsWith('Invalid mac address')) {
+      usage();
+    } else {
+      util.errorHandler(err);
+    }
   }
-
-  // Step1. init bluetoothctl
-  var initPromise = new Promise((resolve, reject) => {
-    bluetoothctl.init((stdout, error) => {
-      if (error) {
-        reject(error);
-      } else {
-        resolve();
-      }
-    });
-  });
-
-  // Step2. try connect to the device
-  initPromise.catch(util.errorHandler);
-  var connectPromise = initPromise.then(() => {
-    return new Promise((resolve, reject) => {
-      connect(mac, util.errorHandler, (isConnected) => {
-        if (isConnected) {
-          resolve(mac);
-        } else {
-          reject(mac);
-        }
-      });
-    });
-  });
-
-  // if succeed, exit
-  connectPromise.then(connectSuccess).then(() => {
-    process.exit();
-  });
-
-  // if fail, scan for another 3 seconds and retry.
-  connectPromise.catch(() => {
-    return new Promise((resolve, reject) => {
-      bluetoothctl.scan(3000, (error) => {
-        if (error) {
-          reject(error);
-        } else {
-          resolve();
-        }
-      });
-    });
-  })
-  .then(() => {
-    connect(mac, util.errorHandler, (isConnected) => {
-      if (isConnected) {
-        connectSuccess(mac);
-      } else {
-        connectFail(mac);
-      }
-    })
-  })
-  .catch(util.errorHandler)
 })(process.argv[2]);
 
 function connectFail(mac) {
@@ -75,26 +25,6 @@ function connectFail(mac) {
 
 function connectSuccess(mac) {
   console.log(mac + ' can be successfully connected.');
-}
-
-function connect(mac, errorCallback, callback) {
-  function isConnected(message) {
-    return message.indexOf('Connection successful') >= 0;
-  }
-
-  bluetoothctl.connect(mac, (stdout, error) => {
-    if (error) {
-      errorCallback(error);
-    } else if (isConnected(stdout)) {
-      callback(true);
-    } else {
-      callback(false);
-    }
-  });
-}
-
-function valid(mac) {
-  return /^([0-9A-Fa-f]{2}\:){5}[0-9A-Fa-f]{2}$/.test(mac);
 }
 
 function usage() {
